Initial commit of exercise 1-13 from the K&R book, not yet complete.
This is a draft of the exercise; at the moment it only prints the length of words.
An idea to proceed further would be to use an array with an uninitialized number of elements.
With that array, each time a new word length is calculated, it is inserted to the array at the correct index.
Finally, after the whole input has been parsed, a histogram will be printed, starting from the lowest word count occurrences to the highest. The character '#' could be used for each occurrence.

For example:
2: ##
3: ####
5: ##
8: #