Updated program to print a histogram relative to a given height.
Instead of drawing a histogram using absolute values, it now uses values relative to the maximum number of occurrences.
For example, the maximum number of occurrences will be represented as the full height of the histogram.
Half that number will have half the height, and so on.

Therefore, the height for each index is now relative, its height is represented as a percentage of the maximum, rounded down.
The percentage obtained is then multiplied with the height define to find out the actual height for that index.
Finally, based on this final value, it is decided what should be drawn for each index.

Due to drawing the histogram based on the macro that defines its height, it will always have the same height.