Updated program to draw the correct number of rows, moved defines to top.
Now program draws the correct number of rows based on the HIST_HEIGHT defined value.
To do this, the first two rows were merged into a single one.
For example, for HIST_HEIGHT with the value 20:
 - the first row used to print the values for 100%
 - the second row used to print the values for 95% - 99.99%
Now they are merged into one, so the first row prints:
 - the values for 95% - 100%

To achieve this the first row (the top of the histogram) is skipped (the if check of the condition: ( i != HIST_HEIGHT )).
Since the following rows include the previous ones, the 2nd row will also include the first.
This was updated in the percentage label column as well.